+++ src/SelfInitializingFakes/XmlFileRecordedCallRepository.cs
@@ -1,4 +1,4 @@
-﻿namespace SelfInitializingFakes
+namespace SelfInitializingFakes
 {
     using System.Collections.Generic;
     using System.IO;
@@ -47,7 +47,7 @@
             using (var reader = XmlReader.Create(fileStream))
             {
 #pragma warning disable CA3075 // Insecure DTD processing in XML - the framework is for testing, so presumably is run in a safe environment
-                return (IEnumerable<RecordedCall>)Serializer.Deserialize(reader);
+                return (IEnumerable<RecordedCall>)Serializer.Deserialize(reader)!;
 #pragma warning restore CA3075 // Insecure DTD processing in XML
             }
         }